refactor: index import stylex
