+++ frontend/apps/web/styles/index.stylex.ts
@@ -1,0 +1,2 @@
+export * from './animations.stylex';
+export * from './colorTokens.stylex';
